pyproject.toml
[project]
name = "richmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.122.0",
    "mysql-connector-python>=9.5.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
